Combine view/edit event
