+++ app/Filament/Resources/EventResource/Pages/EditEvent.php
@@ -2,8 +2,14 @@
 
 namespace App\Filament\Resources\EventResource\Pages;
 
+use App\Filament\Actions\DashboardReport;
 use App\Filament\Resources\EventResource;
+use App\Filament\Resources\EventResource\Widgets\StatsOverview;
+use App\Filament\Resources\EventResource\Widgets\TicketBreakdown;
+use Barryvdh\DomPDF\Facade\Pdf;
 use Filament\Pages\Actions;
+use Filament\Pages\Actions\Action;
+use Filament\Pages\Actions\DeleteAction;
 use Filament\Resources\Pages\EditRecord;
 
 class EditEvent extends EditRecord
@@ -13,8 +19,47 @@ class EditEvent extends EditRecord
     protected function getActions(): array
     {
         return [
-            Actions\ViewAction::make(),
-            Actions\DeleteAction::make(),
+            Action::make('Dashboard Report')
+                ->action(function () {
+                    $breakdown = new TicketBreakdown;
+                    $breakdown->record = $this->record;
+                    $stats = new StatsOverview;
+                    $stats->record = $this->record;
+                    $pdf = Pdf::loadView('pdf.event-dashboard', [
+                        'daysLeft' => $stats->daysLeft,
+                        'reservations' => $stats->reservationTotals,
+                        'orders' => $stats->orderTotals,
+                        'potentialMoney' => $stats->potentialMoney,
+                        'moneyMade' => $stats->moneyMade,
+                        'tablePaid' => $breakdown->tablePaidData(),
+                        'tableFilled' => $breakdown->tableFilledData(),
+                    ]);
+
+                    return response()->streamDownload(
+                        fn () => print($pdf->output()),
+                        'filename.pdf'
+                    );
+                })
+                ->icon('heroicon-s-download'),
+            DeleteAction::make(),
         ];
+    }
+
+    protected function getHeaderWidgetsColumns(): int|array
+    {
+        return 3;
+    }
+
+    protected function getHeaderWidgets(): array
+    {
+        return [
+            StatsOverview::class,
+            TicketBreakdown::class,
+        ];
+    }
+
+    public function hasCombinedRelationManagerTabsWithForm(): bool
+    {
+        return true;
     }
 }
