--- resources/views/filament/resources/events/pages/view-event.blade.php
@@ -1,32 +1,0 @@
-<x-filament::page
-    :widget-data="['record' => $record]"
-    :class="\Illuminate\Support\Arr::toCssClasses([
-        'filament-resources-view-record-page',
-        'filament-resources-' . str_replace('/', '-', $this->getResource()::getSlug()),
-        'filament-resources-record-' . $record->getKey(),
-    ])"
->
-    @php
-        $relationManagers = $this->getRelationManagers();
-    @endphp
-
-    @if (count($relationManagers))
-        @if (! $this->hasCombinedRelationManagerTabsWithForm())
-            <x-filament::hr />
-        @endif
-
-        <x-filament::resources.relation-managers
-            :active-manager="$activeRelationManager"
-            :form-tab-label="$this->getFormTabLabel()"
-            :managers="$relationManagers"
-            :owner-record="$record"
-            :page-class="static::class"
-        >
-            @if ($this->hasCombinedRelationManagerTabsWithForm())
-                <x-slot name="form">
-                    {{ $this->form }}
-                </x-slot>
-            @endif
-        </x-filament::resources.relation-managers>
-    @endif
-</x-filament::page>
